move test_frame_setitem_*
